+++ packages/vm/src/cache.rs
@@ -8,7 +8,7 @@ use crate::backends::{backend, compile};
 use crate::checksum::Checksum;
 use crate::compatibility::check_wasm;
 use crate::errors::{VmError, VmResult};
-use crate::instance::Instance;
+use crate::instance::{Instance, InstanceOptions};
 use crate::modules::FileSystemCache;
 use crate::traits::{Api, Extern, Querier, Storage};
 
@@ -24,7 +24,6 @@ struct Stats {
 pub struct CosmCache<S: Storage, A: Api, Q: Querier> {
     wasm_path: PathBuf,
     supported_features: HashSet<String>,
-    print_debug: bool,
     modules: FileSystemCache,
     stats: Stats,
     // Those two don't store data but only fix type information
@@ -51,7 +50,6 @@ where
     pub unsafe fn new<P: Into<PathBuf>>(
         base_dir: P,
         supported_features: HashSet<String>,
-        print_debug: bool,
     ) -> VmResult<Self> {
         let base = base_dir.into();
         let wasm_path = base.join(WASM_DIR);
@@ -63,7 +61,6 @@ where
         Ok(CosmCache {
             wasm_path,
             supported_features,
-            print_debug,
             modules,
             stats: Stats::default(),
             type_storage: PhantomData::<S>,
@@ -101,19 +98,19 @@ where
         &mut self,
         checksum: &Checksum,
         deps: Extern<S, A, Q>,
-        gas_limit: u64,
+        options: InstanceOptions,
     ) -> VmResult<Instance<S, A, Q>> {
         // try from the module cache
         let res = self.modules.load_with_backend(checksum, backend());
         if let Ok(module) = res {
             self.stats.hits_module += 1;
-            return Instance::from_module(&module, deps, gas_limit, self.print_debug);
+            return Instance::from_module(&module, deps, options.gas_limit, options.print_debug);
         }
 
         // fall back to wasm cache (and re-compiling) - this is for backends that don't support serialization
         let wasm = self.load_wasm(checksum)?;
         self.stats.misses += 1;
-        Instance::from_code(&wasm, deps, gas_limit, self.print_debug)
+        Instance::from_code(&wasm, deps, options)
     }
 }
 
@@ -168,6 +165,10 @@ mod test {
     use wabt::wat2wasm;
 
     const TESTING_GAS_LIMIT: u64 = 400_000;
+    const TESTING_OPTIONS: InstanceOptions = InstanceOptions {
+        gas_limit: TESTING_GAS_LIMIT,
+        print_debug: false,
+    };
     static CONTRACT: &[u8] = include_bytes!("../testdata/contract.wasm");
 
     fn default_features() -> HashSet<String> {
@@ -178,7 +179,7 @@ mod test {
     fn save_wasm_works() {
         let tmp_dir = TempDir::new().unwrap();
         let mut cache: CosmCache<MockStorage, MockApi, MockQuerier> =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
+            unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
         cache.save_wasm(CONTRACT).unwrap();
     }
 
@@ -187,7 +188,7 @@ mod test {
     fn save_wasm_allows_saving_multiple_times() {
         let tmp_dir = TempDir::new().unwrap();
         let mut cache: CosmCache<MockStorage, MockApi, MockQuerier> =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
+            unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
         cache.save_wasm(CONTRACT).unwrap();
         cache.save_wasm(CONTRACT).unwrap();
     }
@@ -208,7 +209,7 @@ mod test {
 
         let tmp_dir = TempDir::new().unwrap();
         let mut cache: CosmCache<MockStorage, MockApi, MockQuerier> =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
+            unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
         let save_result = cache.save_wasm(&wasm);
         match save_result.unwrap_err() {
             VmError::StaticValidationErr { msg, .. } => {
@@ -222,7 +223,7 @@ mod test {
     fn load_wasm_works() {
         let tmp_dir = TempDir::new().unwrap();
         let mut cache: CosmCache<MockStorage, MockApi, MockQuerier> =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
+            unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
 
         let restored = cache.load_wasm(&id).unwrap();
@@ -237,13 +238,13 @@ mod test {
 
         {
             let mut cache1: CosmCache<MockStorage, MockApi, MockQuerier> =
-                unsafe { CosmCache::new(tmp_path, default_features(), false).unwrap() };
+                unsafe { CosmCache::new(tmp_path, default_features()).unwrap() };
             id = cache1.save_wasm(CONTRACT).unwrap();
         }
 
         {
             let cache2: CosmCache<MockStorage, MockApi, MockQuerier> =
-                unsafe { CosmCache::new(tmp_path, default_features(), false).unwrap() };
+                unsafe { CosmCache::new(tmp_path, default_features()).unwrap() };
             let restored = cache2.load_wasm(&id).unwrap();
             assert_eq!(restored, CONTRACT);
         }
@@ -253,7 +254,7 @@ mod test {
     fn load_wasm_errors_for_non_existent_id() {
         let tmp_dir = TempDir::new().unwrap();
         let cache: CosmCache<MockStorage, MockApi, MockQuerier> =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
+            unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
         let checksum = Checksum::from([
             5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5,
             5, 5, 5,
@@ -272,7 +273,7 @@ mod test {
     fn load_wasm_errors_for_corrupted_wasm() {
         let tmp_dir = TempDir::new().unwrap();
         let mut cache: CosmCache<MockStorage, MockApi, MockQuerier> =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
+            unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
         let checksum = cache.save_wasm(CONTRACT).unwrap();
 
         // Corrupt cache file
@@ -291,11 +292,10 @@ mod test {
     #[test]
     fn get_instance_finds_cached_module() {
         let tmp_dir = TempDir::new().unwrap();
-        let mut cache =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
+        let mut cache = unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
         let deps = mock_dependencies(&[]);
-        let _instance = cache.get_instance(&id, deps, TESTING_GAS_LIMIT).unwrap();
+        let _instance = cache.get_instance(&id, deps, TESTING_OPTIONS).unwrap();
         assert_eq!(cache.stats.hits_module, 1);
         assert_eq!(cache.stats.misses, 0);
     }
@@ -303,15 +303,14 @@ mod test {
     #[test]
     fn get_instance_finds_cached_instance() {
         let tmp_dir = TempDir::new().unwrap();
-        let mut cache =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
+        let mut cache = unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
         let deps1 = mock_dependencies(&[]);
         let deps2 = mock_dependencies(&[]);
         let deps3 = mock_dependencies(&[]);
-        let _instance1 = cache.get_instance(&id, deps1, TESTING_GAS_LIMIT).unwrap();
-        let _instance2 = cache.get_instance(&id, deps2, TESTING_GAS_LIMIT).unwrap();
-        let _instance3 = cache.get_instance(&id, deps3, TESTING_GAS_LIMIT).unwrap();
+        let _instance1 = cache.get_instance(&id, deps1, TESTING_OPTIONS).unwrap();
+        let _instance2 = cache.get_instance(&id, deps2, TESTING_OPTIONS).unwrap();
+        let _instance3 = cache.get_instance(&id, deps3, TESTING_OPTIONS).unwrap();
         assert_eq!(cache.stats.hits_module, 3);
         assert_eq!(cache.stats.misses, 0);
     }
@@ -319,11 +318,10 @@ mod test {
     #[test]
     fn init_cached_contract() {
         let tmp_dir = TempDir::new().unwrap();
-        let mut cache =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
+        let mut cache = unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
         let deps = mock_dependencies(&[]);
-        let mut instance = cache.get_instance(&id, deps, TESTING_GAS_LIMIT).unwrap();
+        let mut instance = cache.get_instance(&id, deps, TESTING_OPTIONS).unwrap();
 
         // run contract
         let info = mock_info("creator", &coins(1000, "earth"));
@@ -338,12 +336,11 @@ mod test {
     #[test]
     fn run_cached_contract() {
         let tmp_dir = TempDir::new().unwrap();
-        let mut cache =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
+        let mut cache = unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
         // TODO: contract balance
         let deps = mock_dependencies(&[]);
-        let mut instance = cache.get_instance(&id, deps, TESTING_GAS_LIMIT).unwrap();
+        let mut instance = cache.get_instance(&id, deps, TESTING_OPTIONS).unwrap();
 
         // init contract
         let info = mock_info("creator", &coins(1000, "earth"));
@@ -363,8 +360,7 @@ mod test {
     #[test]
     fn use_multiple_cached_instances_of_same_contract() {
         let tmp_dir = TempDir::new().unwrap();
-        let mut cache =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
+        let mut cache = unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
 
         // these differentiate the two instances of the same contract
@@ -372,7 +368,7 @@ mod test {
         let deps2 = mock_dependencies(&[]);
 
         // init instance 1
-        let mut instance = cache.get_instance(&id, deps1, TESTING_GAS_LIMIT).unwrap();
+        let mut instance = cache.get_instance(&id, deps1, TESTING_OPTIONS).unwrap();
         let info = mock_info("owner1", &coins(1000, "earth"));
         let msg = r#"{"verifier": "sue", "beneficiary": "mary"}"#.as_bytes();
         let res = call_init::<_, _, _, Empty>(&mut instance, &mock_env(), &info, msg).unwrap();
@@ -381,7 +377,7 @@ mod test {
         let deps1 = instance.recycle().unwrap();
 
         // init instance 2
-        let mut instance = cache.get_instance(&id, deps2, TESTING_GAS_LIMIT).unwrap();
+        let mut instance = cache.get_instance(&id, deps2, TESTING_OPTIONS).unwrap();
         let info = mock_info("owner2", &coins(500, "earth"));
         let msg = r#"{"verifier": "bob", "beneficiary": "john"}"#.as_bytes();
         let res = call_init::<_, _, _, Empty>(&mut instance, &mock_env(), &info, msg).unwrap();
@@ -390,7 +386,7 @@ mod test {
         let deps2 = instance.recycle().unwrap();
 
         // run contract 2 - just sanity check - results validate in contract unit tests
-        let mut instance = cache.get_instance(&id, deps2, TESTING_GAS_LIMIT).unwrap();
+        let mut instance = cache.get_instance(&id, deps2, TESTING_OPTIONS).unwrap();
         let info = mock_info("bob", &coins(15, "earth"));
         let msg = br#"{"release":{}}"#;
         let res = call_handle::<_, _, _, Empty>(&mut instance, &mock_env(), &info, msg).unwrap();
@@ -398,7 +394,7 @@ mod test {
         assert_eq!(1, msgs.len());
 
         // run contract 1 - just sanity check - results validate in contract unit tests
-        let mut instance = cache.get_instance(&id, deps1, TESTING_GAS_LIMIT).unwrap();
+        let mut instance = cache.get_instance(&id, deps1, TESTING_OPTIONS).unwrap();
         let info = mock_info("sue", &coins(15, "earth"));
         let msg = br#"{"release":{}}"#;
         let res = call_handle::<_, _, _, Empty>(&mut instance, &mock_env(), &info, msg).unwrap();
@@ -410,15 +406,14 @@ mod test {
     #[cfg(feature = "default-singlepass")]
     fn resets_gas_when_reusing_instance() {
         let tmp_dir = TempDir::new().unwrap();
-        let mut cache =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
+        let mut cache = unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
 
         let deps1 = mock_dependencies(&[]);
         let deps2 = mock_dependencies(&[]);
 
         // Init from module cache
-        let mut instance1 = cache.get_instance(&id, deps1, TESTING_GAS_LIMIT).unwrap();
+        let mut instance1 = cache.get_instance(&id, deps1, TESTING_OPTIONS).unwrap();
         assert_eq!(cache.stats.hits_module, 1);
         assert_eq!(cache.stats.misses, 0);
         let original_gas = instance1.get_gas_left();
@@ -432,7 +427,7 @@ mod test {
         assert!(instance1.get_gas_left() < original_gas);
 
         // Init from instance cache
-        let instance2 = cache.get_instance(&id, deps2, TESTING_GAS_LIMIT).unwrap();
+        let instance2 = cache.get_instance(&id, deps2, TESTING_OPTIONS).unwrap();
         assert_eq!(cache.stats.hits_module, 2);
         assert_eq!(cache.stats.misses, 0);
         assert_eq!(instance2.get_gas_left(), TESTING_GAS_LIMIT);
@@ -442,15 +437,18 @@ mod test {
     #[cfg(feature = "default-singlepass")]
     fn recovers_from_out_of_gas() {
         let tmp_dir = TempDir::new().unwrap();
-        let mut cache =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
+        let mut cache = unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
 
         let deps1 = mock_dependencies(&[]);
         let deps2 = mock_dependencies(&[]);
 
         // Init from module cache
-        let mut instance1 = cache.get_instance(&id, deps1, 10).unwrap();
+        let options = InstanceOptions {
+            gas_limit: 10,
+            print_debug: false,
+        };
+        let mut instance1 = cache.get_instance(&id, deps1, options).unwrap();
         assert_eq!(cache.stats.hits_module, 1);
         assert_eq!(cache.stats.misses, 0);
 
@@ -464,7 +462,11 @@ mod test {
         assert_eq!(instance1.get_gas_left(), 0);
 
         // Init from instance cache
-        let mut instance2 = cache.get_instance(&id, deps2, TESTING_GAS_LIMIT).unwrap();
+        let options = InstanceOptions {
+            gas_limit: TESTING_GAS_LIMIT,
+            print_debug: false,
+        };
+        let mut instance2 = cache.get_instance(&id, deps2, options).unwrap();
         assert_eq!(cache.stats.hits_module, 2);
         assert_eq!(cache.stats.misses, 0);
         assert_eq!(instance2.get_gas_left(), TESTING_GAS_LIMIT);