+++ packages/vm/src/testing/mod.rs
@@ -8,8 +8,9 @@ mod storage;
 
 pub use calls::{handle, init, migrate, query};
 pub use instance::{
-    mock_instance, mock_instance_with_balances, mock_instance_with_failing_api,
-    mock_instance_with_gas_limit, mock_instance_with_options, test_io, MockInstanceOptions,
+    mock_instance, mock_instance_options, mock_instance_with_balances,
+    mock_instance_with_failing_api, mock_instance_with_gas_limit, mock_instance_with_options,
+    test_io, MockInstanceOptions,
 };
 pub use mock::{
     mock_dependencies, mock_dependencies_with_balances, mock_env, mock_info, MockApi,